+++ cider.el
@@ -1259,7 +1259,7 @@ server buffer, in which case a new session for that server is created."
        (cider--update-cljs-init-function)
        (plist-put :session-name ses-name)
        (plist-put :repl-type 'cljs)
-       (plist-put :cider-cljs-repl-pending t)))))
+       (plist-put :cider-repl-cljs-upgrade-pending t)))))
 
 ;;;###autoload
 (defun cider-connect-clj (&optional params)
@@ -1294,7 +1294,7 @@ parameters regardless of their supplied or default values."
      (cider--update-cljs-init-function)
      (plist-put :session-name nil)
      (plist-put :repl-type 'cljs)
-     (plist-put :cider-cljs-repl-pending t))))
+     (plist-put :cider-repl-cljs-upgrade-pending t))))
 
 ;;;###autoload
 (defun cider-connect-clj&cljs (params &optional soft-cljs-start)